% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.r
\name{plotStates}
\alias{plotStates}
\title{Plot state variables}
\usage{
plotStates(out, out_old, timeUnit, timeBase, model, mult, show, rangeT, rangeY,
  gridT, gridY, logY, showOld, obs)
}
\arguments{
\item{out}{Object returned from latest call to \code{simul}.}

\item{out_old}{Object returned from an earlier call to \code{simul}.}

\item{timeUnit}{Unit of numeric simulation time info (as a string).}

\item{timeBase}{Date and time of the startpoint of the simulation as a
value of class \code{POSIXct}.}

\item{model}{Object of class \code{rodeo} representing the model.}

\item{mult}{Named vector with multipliers to apply to variables before plotting.}

\item{show}{Named logical vector to enable plotting of individual variables.}

\item{rangeT}{Range of the x-axis (time axis).}

\item{rangeY}{Range of the y-axis.}

\item{gridT}{Logical to enable/disable grid for x-axis (time axis).}

\item{gridY}{Logical to enable/disable grid for y-axis.}

\item{logY}{Logical to switch between linear and log axis.}

\item{showOld}{Logical. Enables/disables plotting of data in \code{out_old}.}

\item{obs}{\code{NULL} or a data frame with observations (numeric time info
  is expected in the first column).}
}
\value{
\code{NULL}.
}
\description{
Plots the simulated dynamics of state variables.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

