% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{prettyTimes}
\alias{prettyTimes}
\title{Create a vector of times for labelling a time axis.}
\usage{
prettyTimes(tmin, tmax, n)
}
\arguments{
\item{tmin}{Lower limit of time axis (\code{POSIXct}).}

\item{tmax}{Upper limit of time axis (\code{POSIXct}).}

\item{n}{Desired number of axis ticks.}
}
\value{
A list with components \code{values} and \code{format}. See example
  for usage.
}
\description{
This algorithm usually gives better results than axis.POSIXt.
}
\examples{
\dontrun{
  times= prettyTimes(lower, upper, 10)
  axis.POSIXct(1, at=times$values, format=times$format, las=2, xlab="")
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

