% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_rodeoApp.r
\docType{package}
\name{rodeoApp-package}
\alias{rodeoApp}
\alias{rodeoApp-package}
\title{Interfaces for rodeo-based models}
\description{
This package provides convenient interfaces for rodeo-based models. In
particular, it allows for running a rodeo-based from a (shiny) graphical user
interface (see \code{\link{runGUI}}). Type \code{help(package="rodeoApp")} to
inspect the package description.
}

