% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runable.r
\name{runGUI}
\alias{runGUI}
\title{Run rodeo-based model in shiny GUI}
\usage{
runGUI(dir = "", xlFile = "model.xlsx", funsR = "functions.r",
  funsF = "functions.f95", tables = c(vars = "vars", pars = "pars", funs =
  "funs", pros = "pros", stoi = "stoi"), colsep = ",", obs = NULL,
  dllname = NULL, serverMode = FALSE)
}
\arguments{
\item{dir}{Directory where input files are located. This value is added as a
prefix to any file names (i.e. \code{xlFile}, \code{funsR}, \code{funsF},
and possibly the values of \code{tables}).}

\item{xlFile}{Either the base name of file in MS Excel format holding
the model definition or \code{NULL}. In the latter case, the model
definition is read from delimited text files using
\code{read.table(header=TRUE, sep=colsep)} where the last argument is
under user control (see \code{colsep} below).}

\item{funsR}{Base name of file with function definitions in R.}

\item{funsF}{Base name of file with function definitions in Fortran 95.}

\item{tables}{Named vector with required elements 'vars', 'pars', 'funs',
'pros', and 'stoi'. If a file name was supplied in \code{xlFile}, the
values are interpreted as the names of worksheets holding the declaration
of variables, parameters, functions, process rates, and stoichiometry
factors, respectively. If \code{xlFile} is an empty string, the values are
interpreted as the names of delimited text files holding the respective
information.}

\item{colsep}{Specifies the column separator in text files. Only used if
\code{xlFile} is \code{NULL}.}

\item{obs}{If not \code{NULL}, this must be a data frame with observed values
of the state variables. Time information (numeric) is expected in the first
column. Columns with observation data (if existing) must be named like the
simulated state variables. Missing values must be marked as \code{NA}.}

\item{dllname}{Character string used as the name for the built library. This
  is the base name without file extension (must not contain path separators,
  white space, etc.). If set to \code{NULL} (the default), an automatic
  name is assigned. The library is created in the folder returned by
  \code{tempdir()} and it carries the system-specific extension for shared
  libraries according to \code{.Platform$dynlib.ext}.}

\item{serverMode}{Defaults to \code{FALSE}. If set to \code{TRUE}, data
  required by the GUI are saved to disk but the shiny app is
  \emph{not} run.}
}
\value{
If \code{serverMode} is \code{FALSE}, the function returns
  \code{NULL}. Otherwise it returns a vector holding the names of 3
  files needed to run the GUI. These files carry the extension .r, .rda, and
  .so or .dll, respectively. They reside in the current working directory.
}
\description{
Runs a rodeo-based model in a shiny GUI to allow for variation of parameters
and/or initial values.
}
\note{
Existing files are not overwritten unless they reside in R's temporary
  folder. Thus, manual file deletion may be necessary when \code{serverMode}
  is \code{TRUE}.
}
\examples{
\dontrun{
  library(rodeoApp)
  runGUI(dir=system.file("examples/DRT", package="rodeoApp"))
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

