% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runable.r
\name{runMCS}
\alias{runMCS}
\title{Run rodeo-based model in Monte Carlo mode}
\usage{
runMCS(outdir, times, nruns, nsig = 4, overwrite = FALSE, dir = "",
  xlFile = "model.xlsx", funsR = "functions.r", funsF = "functions.f95",
  sheets = c(vars = "vars", pars = "pars", funs = "funs", pros = "pros", stoi
  = "stoi"), silent = FALSE)
}
\arguments{
\item{outdir}{Output directory. Should be empty unless \code{overwrite} is \code{TRUE}.}

\item{times}{Vector of times for which the states are computed.}

\item{nruns}{Desired number of model runs.}

\item{nsig}{Number of significant digits in output files. Should be set to
a reasonably small value for large \code{nruns} in order to save disk space.}

\item{overwrite}{Logical. Is it OK to overwrite existing output files?}

\item{dir}{Directory where input files are located. This value is added as a
prefix to any file names (i.e. \code{xlFile}, \code{funsR}, \code{funsF}).}

\item{xlFile}{Base name of file in MS Excel format with the model definition.}

\item{funsR}{Base name of file with function definitions in R.}

\item{funsF}{Base name of file with function definitions in Fortran 95.}

\item{sheets}{Named vector with required elements 'vars', 'pars', 'funs',
  'pros', and 'stoi'. These elements define the names of the worksheets in
  \code{xlFile} holding the declaration of variables, parameters, functions,
  process rates, and stoichiometry factors, respectively.}

\item{silent}{Logical. Used to switch diagnostic messages on/off.}
}
\value{
\code{NULL}, see notes
}
\description{
Runs a rodeo-based model in Monte Carlo mode with parameters and initial
values sampled from uniform distributions (latin hypercube method).
}
\note{
The function creates three TAB-separated text files in \code{outdir}:
\itemize{
  \item{\code{var.txt} : } Initial values of state variables (one record for
    each run).
  \item{\code{par.txt} : } Parameters (one record for each run).
  \item{\code{sim.txt} : } Simulation results. For each run, the number of
    records equals the number of elements in \code{times}.
}
The first column in all these tables is named 'runID'. It is the primary
key column to be used when merging the tables, e.g. using \code{\link{merge}}.
}
\examples{
\dontrun{
  library(rodeoApp)
  runMCS(
    outdir=gsub(pattern="\\\\", replacement="/", x=tempdir(), fixed=TRUE),
    times=seq(0,30,1/24), nruns=10, nsig=4, overwrite=FALSE,
    dir=system.file("examples/DRT", package="rodeoApp"))
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

