% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.r
\name{simul}
\alias{simul}
\title{Run a dynamic simulation}
\usage{
simul(model, vars, pars, times, dllfile)
}
\arguments{
\item{model}{Object of class \code{rodeo} representing the model.}

\item{vars}{Named vector of the state variables' initial values.}

\item{pars}{Named vector of parameters.}

\item{times}{Vector of times for which the states are computed.}

\item{dllfile}{Shared library file holding the compiled model.}
}
\value{
The object returned by \code{deSolve::ode}.
}
\description{
Computes the values of the state variables for a sequence of times.
}
\note{
An error is generated if the integration was not successful.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

