% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simul.r
\name{stst}
\alias{stst}
\title{Compute steady-state solution}
\usage{
stst(model, vars, pars, dllfile, rtol, atol)
}
\arguments{
\item{model}{Object of class \code{rodeo} representing the model.}

\item{vars}{Named vector of the state variables' initial values.}

\item{pars}{Named vector of parameters.}

\item{dllfile}{Shared library file holding the compiled model.}

\item{rtol}{Vector of the same length as \code{vars} specifying the
relative tolerance for the solver (see help of \code{deSolve::lsoda}).}

\item{atol}{Vector of the same length as \code{vars} specifying the
  absolute tolerance for the solver (see help of \code{deSolve::lsoda}).}
}
\value{
The object returned by \code{rootSolve::steady}. The \code{y}-
  component of this object has names based on the \code{ynames}
  attribute.
}
\description{
Estimates the values of the state variables for steady-state conditions.
}
\note{
An error is generated if steady-state estimation was not successful.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

