% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.r
\name{timeFmt}
\alias{timeFmt}
\title{Find a reasonable time format}
\usage{
timeFmt(tmin, tmax)
}
\arguments{
\item{tmin}{Lower limit of time axis (\code{POSIXct}).}

\item{tmax}{Upper limit of time axis (\code{POSIXct}).}
}
\value{
A format string.
}
\description{
Finds an appropriate time format specifier for a given time range
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

