% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{visStoi}
\alias{visStoi}
\title{Visualize the stoichiometry matrix}
\usage{
visStoi(model, time, vars, pars, funsR)
}
\arguments{
\item{model}{Object of class \code{rodeo} representing the model.}

\item{time}{Time at which stoichiometry is to be evaluated.}

\item{vars}{Named vector of the state variables' initial values.}

\item{pars}{Named vector of parameters.}

\item{funsR}{File with function definitions as R code.}
}
\value{
Name of a temporary file containing generated HTML code.
}
\description{
Creates a graphical representation of the stoichiometry matrix.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

