% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.r
\name{plotStates}
\alias{plotStates}
\title{Plot state variables}
\usage{
plotStates(out, out_old, model, draw, yrange, showOld)
}
\arguments{
\item{out}{Object returned from latest call to \code{simul}.}

\item{out_old}{Object returned from an earlier call to \code{simul}.}

\item{model}{Object of class \code{rodeo} representing the model.}

\item{draw}{Named logical vector to enable plotting of individual variables.}

\item{yrange}{Range of the y-axis.}

\item{showOld}{Logical. Enables/disables plotting of data in \code{out_old}.}
}
\value{
\code{NULL}.
}
\description{
Plots the simulated dynamics of state variables.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

