% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_rodeoApp.r
\docType{package}
\name{rodeoApp-package}
\alias{rodeoApp-package}
\title{Run a rodeo-based model as a shiny application}
\description{
Type \code{help(package="rat")} to inspect the package description.
}

