% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runable.r
\name{runGUI}
\alias{runGUI}
\title{Run rodeo-based model in shiny GUI}
\usage{
runGUI(dir = "", xlFile = "model.xlsx", funsR = "functions.r",
  funsF = "functions.f95", tables = c(vars = "vars", pars = "pars", funs =
  "funs", pros = "pros", stoi = "stoi"), colsep = ";")
}
\arguments{
\item{dir}{Directory where input files are located. This value is added as a
prefix to any file names (i.e. \code{xlFile}, \code{funsR}, \code{funsF},
and possibly the values of \code{tables}).}

\item{xlFile}{Either the base name of file in MS Excel format holding
the model definition or \code{NULL}. In the latter case, the model
definition is read from delimited text files.}

\item{funsR}{Base name of file with function definitions in R.}

\item{funsF}{Base name of file with function definitions in Fortran 95.}

\item{tables}{Named vector with required elements 'vars', 'pars', 'funs',
'pros', and 'stoi'. If a file name was supplied in \code{xlFile}, the
values are interpreted as the names of worksheets holding the declaration
of variables, parameters, functions, process rates, and stoichiometry
factors, respectively. If \code{xlFile} is an empty string, the values are
interpreted as the names of delimited text files holding the respective
information.}

\item{colsep}{Specifies the column separator in text files. Only used if
  \code{xlFile} is \code{NULL}.}
}
\value{
\code{NULL}
}
\description{
Runs a rodeo-based model in a shiny GUI to allow for variation of parameters
and/or initial values.
}
\note{
The function exports variables to the global environment for
  use by the shiny server and user interface.
}
\examples{
\dontrun{
  library(rodeoApp)
  runGUI(dir=system.file("examples/DRT", package="rodeoApp"))
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

