% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runable.r
\name{runGUI}
\alias{runGUI}
\title{Run rodeo-based model in shiny GUI}
\usage{
runGUI(dir = "", xlFile = "model.xlsx", funsR = "functions.r",
  funsF = "functions.f95", sheets = c(vars = "vars", pars = "pars", funs =
  "funs", pros = "pros", stoi = "stoi"))
}
\arguments{
\item{dir}{Directory where input files are located. This value is added as a
prefix to any file names (i.e. \code{xlFile}, \code{funsR}, \code{funsF}).}

\item{xlFile}{Base name of file in MS Excel format with the model definition.}

\item{funsR}{Base name of file with function definitions in R.}

\item{funsF}{Base name of file with function definitions in Fortran 95.}

\item{sheets}{Named vector with required elements 'vars', 'pars', 'funs',
  'pros', and 'stoi'. These elements define the names of the worksheets in
  \code{xlFile} holding the declaration of variables, parameters, functions,
  process rates, and stoichiometry factors, respectively.}
}
\value{
\code{NULL}
}
\description{
Runs a rodeo-based model in a shiny GUI to allow for variation of parameters
and/or initial values.
}
\note{
The function exports variables to the global environment for
  use by the shiny server and user interface.
}
\examples{
\dontrun{
  library(rodeoApp)
  runGUI(dir=system.file("examples/DRT", package="rodeoApp"))
}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

